fix(minjae/develop): update page 유효성 검사
