+++ app/pages/boards/[categoryName]/[num]/update.js
@@ -1,11 +1,29 @@
-import React from "react";
+import React, { useEffect } from "react";
 import BoardBanner from "../../../../components/Board/BoardBanner";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import UpdateEditor from "../../../../components/Editor/UpdateEdtior";
+import { useSelector } from "react-redux";
 
 const BoardUpdatePage = () => {
   const router = useRouter();
+  const { categoryName, num } = router.query;
+
+  const { id } = useSelector((state) => state.auth);
+  const board = useSelector((state) => state.board);
+
+  useEffect(() => {
+    if (id.length > 0 && board.studentId.length > 0) {
+      if (board.studentId !== id) {
+        alert("잘못된 접근 방식입니다.");
+        router.back();
+      }
+    } else if (board.studentId.length === 0) {
+      alert("URL을 통한 접근을 금지합니다.");
+      router.back();
+    }
+  }, []);
+
   return (
     <>
       <Head>
@@ -13,7 +31,7 @@ const BoardUpdatePage = () => {
       </Head>
       <BoardBanner
         title={
-          router.query === "notice" || router.query === "free"
+          categoryName === "notice" || categoryName === "free"
             ? "boards"
             : "Market"
         }
@@ -21,7 +39,12 @@ const BoardUpdatePage = () => {
       />
       <section id="post-write" className="post-write">
         <div className="container">
-          <UpdateEditor />
+          <UpdateEditor
+            categoryName={categoryName}
+            num={num}
+            board={board}
+            id={id}
+          />
         </div>
       </section>
     </>
